+++ src/esm/components/form/text-input-with-icon/index.js
@@ -6,13 +6,12 @@ import styled from 'styled-components';
 import { pxToRem } from '../../../helpers/utils/typography';
 import { TextInput } from '../../../components/form/text-input';
 import PropTypes from 'prop-types';
-import COLORS from '../../../constants/colors-config';
 import { VisuallyHidden } from '../../accessibility/visually-hidden';
 import classNames from 'classnames';
 var StyledTextInputWithIcon = styled.div.withConfig({
   displayName: "text-input-with-icon__StyledTextInputWithIcon",
   componentId: "sc-rvt8kr-0"
-})(["position:relative;.k-Form-TextInputWithIcon__icon{display:flex;position:absolute;align-items:center;justify-content:center;z-index:1;top:0;width:", ";height:100%;&.k-Form-TextInputWithIcon__icon--disabled > svg:not(.k-ColorSvg){&[stroke]:not([stroke='none']),& [stroke]:not([stroke='none']){stroke:", ";}&[fill]:not([fill='none']),& [fill]:not([fill='none']){fill:", ";}}}&.k-Form-TextInputWithIcon--icon_left{.k-Form-TextInputWithIcon__input{padding-left:", ";}.k-Form-TextInputWithIcon__icon{left:0;}}&.k-Form-TextInputWithIcon--icon_right{.k-Form-TextInputWithIcon__input{padding-right:", ";}.k-Form-TextInputWithIcon__icon{right:0;}}"], pxToRem(50), COLORS.font2, COLORS.font2, pxToRem(50), pxToRem(50));
+})(["position:relative;.k-Form-TextInputWithIcon__icon{display:flex;position:absolute;align-items:center;justify-content:center;z-index:1;top:0;width:", ";height:100%;&.k-Form-TextInputWithIcon__icon--disabled > svg:not(.k-ColorSvg){&[stroke]:not([stroke='none']),& [stroke]:not([stroke='none']){stroke:var(--color-grey-500);}&[fill]:not([fill='none']),& [fill]:not([fill='none']){fill:var(--color-grey-500);}}}&.k-Form-TextInputWithIcon--icon_left{.k-Form-TextInputWithIcon__input{padding-left:", ";}.k-Form-TextInputWithIcon__icon{left:0;}}&.k-Form-TextInputWithIcon--icon_right{.k-Form-TextInputWithIcon__input{padding-right:", ";}.k-Form-TextInputWithIcon__icon{right:0;}}"], pxToRem(50), pxToRem(50), pxToRem(50));
 export var TextInputWithIcon = function TextInputWithIcon(_ref) {
   var disabled = _ref.disabled,
       icon = _ref.icon,