+++ src/esm/components/graphics/logos/youmatter/index.js
@@ -50,5 +50,5 @@ YouMatterLogo.propTypes = {
 YouMatterLogo.defaultProps = {
   width: null,
   height: null,
-  color: "#222"
+  color: '#222'
 };